--- a/.github/workflows/gocover.yml
+++ b/.github/workflows/gocover.yml
@@ -7,15 +7,20 @@ on:
     branches: [ main ]
   workflow_dispatch:
 
+permissions:
+    contents: read
+
 jobs:
 
   build:
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/checkout@v3
+    - uses: actions/checkout@v5
+      with:
+        persist-credentials: false
 
     - name: Set up Go
-      uses: actions/setup-go@v3
+      uses: actions/setup-go@v6
       with:
         go-version: stable
 
@@ -23,7 +28,7 @@ jobs:
       run: go build -v ./...
 
     - name: Test
-      run: go test -v ./...
+      run: go test -race -v ./...
 
     - name: Update coverage badge
-      uses: ncruces/go-coverage-report@main
+      uses: ncruces/go-coverage-report@v0