--- a/.github/workflows/gocover.yml
+++ b/.github/workflows/gocover.yml
@@ -2,9 +2,9 @@ name: gocover
 
 on:
   push:
-    branches: [ main ]
+    branches: [ master ]
   pull_request:
-    branches: [ main ]
+    branches: [ master ]
   workflow_dispatch:
 
 permissions: