--- a/go.mod
+++ b/go.mod
@@ -1,6 +1,6 @@
 module github.com/kevburnsjr/skipfilter
 
-go 1.16
+go 1.18
 
 require (
 	github.com/MauriceGit/skiplist v0.0.0-20191117202105-643e379adb62