--- a/go.mod
+++ b/go.mod
@@ -3,7 +3,12 @@ module github.com/kevburnsjr/skipfilter
 go 1.18
 
 require (
-	github.com/MauriceGit/skiplist v0.0.0-20191117202105-643e379adb62
-	github.com/RoaringBitmap/roaring v0.9.4
-	github.com/hashicorp/golang-lru v0.5.4
+	github.com/MauriceGit/skiplist v0.0.0-20211105230623-77f5c8d3e145
+	github.com/RoaringBitmap/roaring v1.9.4
+	github.com/hashicorp/golang-lru/v2 v2.0.7
+)
+
+require (
+	github.com/bits-and-blooms/bitset v1.24.0 // indirect
+	github.com/mschoch/smat v0.2.0 // indirect
 )