--- a/go.mod
+++ b/go.mod
@@ -1,14 +1,15 @@
 module github.com/kevburnsjr/skipfilter
 
-go 1.18
+go 1.24.0
 
 require (
 	github.com/MauriceGit/skiplist v0.0.0-20211105230623-77f5c8d3e145
 	github.com/RoaringBitmap/roaring/v2 v2.10.0
-	github.com/hashicorp/golang-lru/v2 v2.0.7
+	github.com/maypok86/otter/v2 v2.2.1
 )
 
 require (
 	github.com/bits-and-blooms/bitset v1.24.0 // indirect
 	github.com/mschoch/smat v0.2.0 // indirect
+	golang.org/x/sys v0.36.0 // indirect
 )