Switch from golang-lru to Otter, which is known to be more performant.

I added some benchmarks, the diff shows a significant improvement:

```
goarch: arm64
pkg: github.com/kevburnsjr/skipfilter
cpu: Apple M1 Pro
                                 │    old.txt    │               new.txt                │
                                 │    sec/op     │    sec/op     vs base                │
SkipFilter_MatchAny_Read-10        1017.6µ ± 15%   326.2µ ±  1%  -67.95% (p=0.000 n=10)
SkipFilter_MatchAny_ReadWrite-10   1664.7µ ±  2%   893.3µ ± 12%  -46.34% (p=0.000 n=10)
geomean                             1.302m         539.8µ        -58.53%
```